fast_api/requirements.txt
aiofiles==23.2.1
fastapi==0.110.2
jsonpickle==3.0.4
numpy==1.26.4
pdf2image==1.17.0
Pillow==10.3.0
pydantic==2.7.1
pytesseract==0.3.10
spacy==3.7.4
spacy_transformers==1.3.4
SQLAlchemy==2.0.29
surya==1.0.0
torch==2.2.2
ultralytics==8.2.2
